test: pagination with preview header
